Add tests for verify functions
* Add tests for verify_actions, verify_reasons and verify_token
* Move verify_actions and verify_reasons to helpers.py
