+++ timereport/api.py
@@ -3,11 +3,10 @@ import json
 import logging
 
 from timereport.lib.factory import factory, date_to_string, json_serial
-from timereport.lib.slack import slack_payload_extractor, verify_token, verify_actions, verify_reasons
+from timereport.lib.slack import slack_payload_extractor, verify_token
 from timereport.lib.add import post_to_backend, create_event
 from timereport.lib.list import get_between_date, get_user_by_id
-from timereport.lib.helpers import parse_config
-
+from timereport.lib.helpers import parse_config, verify_reasons, verify_actions
 
 logger = logging.getLogger()
